close to axios call functionality for both speech and text, text working at this time
