+++ client/src/App.css
@@ -43,7 +43,7 @@
 }
 
 .feedback-form{
-  background-color: #d7ddff;
+  /* background-color: #90caf9; */
   border-radius: 10px;
   width: 44vw;
   margin-left: 20vw;
@@ -54,14 +54,14 @@
   width: 40vw;
   margin-left: 2vw;
   margin-bottom: 2vw;
-  border: none;
+  border: 1px solid #90caf9;
   border-radius: 10px;
   font-family: Arial, Helvetica, sans-serif;
 }
 
 .feedback-buttons-row{
   width: 44vw;
-  padding: 0.5em;
+  padding-bottom: 0.5em;
   /* text-align: center; */
   /* border: 1px solid black; */
   display: flex;
@@ -69,7 +69,6 @@
 }
 
 .feedback-instructions{
-  padding: 1em;
   /* border: 1px solid black; */
   display: flex;
   justify-content: center;
